Unnecesary UC flags removed
As detailed in this document: https://www.dte.us.es/docencia/etsii/gii-is/estructura-de-computadores/grupo-4-2018/mcs2010, the CS2010's UC implementation could send some "unnecesary" signals in order to make easier the design. However, as the Logisim implementation uses a ROM-based UC, this trick has never been useful for it. The UC's ROM has been updated removing the unnecesary signals for each instruction, expecting it to be clearer when following the UC procedures.